move version 2 to default, set wires differently
